[common] added resample support
While this on the surface may not make much sense to do, according to Xiph the
RnNoise neural net is computed specifically for 48KHz and as such would require
a re-train simply to operate at a different sample rate.

This patch also implements a ring buffer to avoid excessive memory operations.