CommonPlugin: Add initial support for VAD
rnnoise can return the probability of a chunk being voice when
processed. We can use this probability to define a threshold under which
we return silence, giving us VAD (Voice Activity Detection)
capabilities. This is an intended usecase of rnnoise.

Currently this threshold is hardcoded in a #define. The different
plugins (LADSPA, VST, LV2) will need to be expanded with the respective
configuration APIs and then this value can be passed to the
CommonPlugin.