CommonPlugin: Add grace period to VAD
This adds a grace period to the voice activity detection. After a frame
was identified as speech, we allow N samples after it to pass through
the VAD clamp, irregardless of rnnoise' classification.

Without this we can have cut outs in the middle of words which, while
intelligible, make the voice quality sound bad.