Use JUCE, new parameters, stereo, tests
This is a huge BREAKING refactoring. If you load LADSPA plugin with
pulseaudio - it will be BROKEN because new parameters are added.

JUCE has GPLv3 license so we could use it and change however we want.

JUCE brings us the following benefits:
- We could support all plugins which JUCE supports without any
  additional code for each plugin. Now supported via JUCE:
  - all systems: LV2 VST VST3
  - macOS only: AU AUv3
- Better compatibility with different plugin hosts.
- Easy to use FST headers instead of VST2. FST is an open-source
  reverse-engineered headers for VST2; they are needed because
  VST2 license does not permit its usage.
- GUI!

Downsides are:
- Longer compile times;
- Bigger library sizes;
- Have to vendor JUCE (via git subproject) in order to patch it.

- New parameter: retroactive VAD grace period - it introduces some
  latency and in return allows to fix a small cut off at sentence start.
- Reworked block processing to be more modular/readable and support
  blocks with variable frames count (e.g. used by Audacity).

Tests are implemented via catch2. Tests aren't checking much, but should
be enough to check undefined behavior and memory corruption.

Build artifacts for Windows, Linux, macOS on CI.